DESeq multi factor wizard: Bug fixes when going backwards in panels. Simple Test Plotting: You can now click on a point in the plot and the viewer will jump to the position of this point. ToolTips are included.

+++ VAMP_NBP/VAMP_Exporter/src/de/cebitec/vamp/exporter/excel/ExcelExporter.java
@@ -231,7 +231,7 @@ public class ExcelExporter {
             sheet.addCell(label);
         } else if (celltype.equals("INTEGER")) {
             WritableCellFormat integerFormat = new WritableCellFormat(NumberFormats.INTEGER);
-            Double value = Double.parseDouble(cellvalue.toString());
+            Integer value = Integer.parseInt(cellvalue.toString());
             Number number = new Number(column, row, value, integerFormat);
             sheet.addCell(number);
         } else if (celltype.equals("DOUBLE")) {
